library-backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.1
PyJWT==2.8.0
PyMySQL==1.1.1
python-dotenv==1.0.1
marshmallow==3.21.3
cryptography==42.0.8
